feature : adding viewmodel and di #PATRON2020-37
