+++ app/src/main/java/com/intive/patronage/smarthome/di/DashboardModule.kt
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import com.intive.patronage.smarthome.feature.dashboard.model.DashboardSensor
 import com.intive.patronage.smarthome.feature.dashboard.view.SensorsListAdapter
 import com.intive.patronage.smarthome.feature.dashboard.viewmodel.DashboardViewModel
+import com.intive.patronage.smarthome.feature.hvac.HvacViewModel
 import com.intive.patronage.smarthome.navigator.DashboardCoordinator
 import com.intive.patronage.smarthome.navigator.Navigator
 import org.koin.dsl.module
