+++ app/src/main/res/layout-land/fragment_hvac_details.xml
@@ -3,17 +3,23 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <data>
+        <variable name="hvacViewModel"
+                  type="com.intive.patronage.smarthome.feature.hvac.HvacViewModel"/>
+    </data>
 
     <LinearLayout
             android:background="@color/backgroundColor"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+
+    >
 
         <com.intive.patronage.smarthome.feature.hvac.HvacCircle
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
                 android:layout_weight="4"
-                android:id="@+id/havacCircle"
+                android:id="@+id/hvac_circle"
         />
         <LinearLayout
                 android:layout_width="0dp"
