+++ app/src/main/res/layout/fragment_hvac_details.xml
@@ -2,7 +2,10 @@
 <layout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto">
-
+    <data>
+        <variable name="hvacViewModel"
+                  type="com.intive.patronage.smarthome.feature.hvac.HvacViewModel"/>
+    </data>
 
     <LinearLayout
             android:background="@color/backgroundColor"
@@ -14,7 +17,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_weight="4"
-                android:id="@+id/havacCircle"
+                android:id="@+id/hvac_circle"
+
                 />
         <LinearLayout
                 android:layout_width="match_parent"
